requirements.txt
alembic==1.7.5
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
docopt==0.6.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pipreqs==0.4.11
postgres==4.0
psycopg2==2.9.2
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
